+++ .idea/deploymentTargetSelector.xml
@@ -4,10 +4,10 @@
     <selectionStates>
       <SelectionState runConfigName="app">
         <option name="selectionMode" value="DROPDOWN" />
-        <DropdownSelection timestamp="2024-09-30T00:10:22.193059Z">
+        <DropdownSelection timestamp="2024-10-17T16:07:08.332171900Z">
           <Target type="DEFAULT_BOOT">
             <handle>
-              <DeviceId pluginId="LocalEmulator" identifier="path=/Users/alexholt/.android/avd/Nexus_4_API_24.avd" />
+              <DeviceId pluginId="LocalEmulator" identifier="path=C:\Users\bp\.android\avd\Pixel_API_24.avd" />
             </handle>
           </Target>
         </DropdownSelection>
