+++ app/src/main/java/com/example/fitfeed/MainActivity.java
@@ -11,6 +11,9 @@ import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.Fragment;
 
+import com.example.fitfeed.fragments.HomeFragment;
+import com.example.fitfeed.fragments.SocialFragment;
+import com.example.fitfeed.fragments.WorkoutsFragment;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 
 public class MainActivity extends AppCompatActivity {
