+++ app/src/main/java/com/example/fitfeed/fragments/HomeFragment.java
@@ -1,4 +1,4 @@
-package com.example.fitfeed;
+package com.example.fitfeed.fragments;
 
 import android.os.Bundle;
 
@@ -8,12 +8,14 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.example.fitfeed.R;
+
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link WorkoutsFragment#newInstance} factory method to
+ * Use the {@link HomeFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class WorkoutsFragment extends Fragment {
+public class HomeFragment extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -24,7 +26,7 @@ public class WorkoutsFragment extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public WorkoutsFragment() {
+    public HomeFragment() {
         // Required empty public constructor
     }
 
@@ -34,11 +36,11 @@ public class WorkoutsFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment WorkoutsFragment.
+     * @return A new instance of fragment HomeFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static WorkoutsFragment newInstance(String param1, String param2) {
-        WorkoutsFragment fragment = new WorkoutsFragment();
+    public static HomeFragment newInstance(String param1, String param2) {
+        HomeFragment fragment = new HomeFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -59,6 +61,6 @@ public class WorkoutsFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_workouts, container, false);
+        return inflater.inflate(R.layout.fragment_home, container, false);
     }
 }
