+++ app/src/main/java/com/example/fitfeed/fragments/SocialFragment.java
@@ -1,4 +1,4 @@
-package com.example.fitfeed;
+package com.example.fitfeed.fragments;
 
 import android.content.Intent;
 import android.graphics.drawable.Drawable;
@@ -15,6 +15,9 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.example.fitfeed.FriendsActivity;
+import com.example.fitfeed.R;
+import com.example.fitfeed.adapters.PostsRecyclerViewAdapter;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 
 import java.util.ArrayList;
