+++ app/src/main/res/layout-land/fragment_social.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/social_bg"
-    tools:context=".HomeFragment">
+    tools:context=".fragments.HomeFragment">
 
     <com.google.android.material.floatingactionbutton.FloatingActionButton
         android:id="@+id/addFriendsButton"
