+++ app/src/main/res/layout/fragment_workouts.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HomeFragment"
+    tools:context=".fragments.HomeFragment"
     android:background="@color/workouts_bg">
 
     <TextView
